requirements.txt
asgiref==3.5.1
astroid==2.11.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.4
Django==4.0.4
django-allauth==0.50.0
django-crispy-forms==1.14.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
